requirements.txt
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
wheel
gunicorn
Flask-OAuthlib
Flask-PyMongo
bson